Added more functions for cbs.

+++ competitors/include/node.h
@@ -6,12 +6,12 @@
 struct Node {
     Node* parent;
     Location location;
-    int g;     // only +-1 by step
+    unsigned g;     // only +-1 by step
     double h;
     double f;
 
     Node(Location location): parent(nullptr), location(location), g(0), h(0), f(0) {}
-    Node(Node* parent, Location location, double g, double h): parent(parent), location(location), g(g), h(h), f(g+h) {}
+    Node(Node* parent, Location location, unsigned g, double h): parent(parent), location(location), g(g), h(h), f(g+h) {}
     ~Node() {};
 
     bool operator < (const Node* other) const {
